layouts/page: add note on usage for the about page
The previous layout was informed by discussion in the forums [1]
although it is difficult to find specific concrete documentation on the
behavior that seems to have a pagic "layouts/page/" directory in the
lookup order.

This new scheme is meant to be clearer, with the directory name clearly
specifying the relationship between the about page content and it's HTML
representation.

[1] https://discourse.gohugo.io/t/how-to-add-an-extra-standalone-page-on-single-page-hugo-icon-theme/9661/6